Lot Done
Data Base:
- Did most of the data base still needs a quick rewrite to the tasks to
make a       linker table tasks. This would be useful in the way of
linking multiple tasks to each other vs 0-1 there is now

WebApp
-Lots here. Most of it has to do with the accounts as I took my time to
do that in an impactful way. such as salting and hashing passwords.
Custom athorization with a data base first style. I also made some
custom extentions/hellper for both the controller and razor systems.
->cs Files:
-CustomWebHelpersAndExtentions : a few classes here
-Password helper tool : a static class that can be set to set string
extention methods for password checks
- CustomWebExtentions : Has some string and HtmlHelper extentions you
may find useful. Such as a easy and easy to read script/style bundle
adding that fixes bundle errors by ensuring head placement from any page
or controller.
- CustomContollerExtentions : some conroller extentions such as a bundle
injection method to be used to add to the page. Wouldn't recomend it as
the view can handle it but it is there.
- CustomControllerHelper : some controller helpers so you can do things
quickly
- CustomWebHelper : simple web helpers

-DBFSPAuthorizeAttribute : Has a class by the same name that Authorizes
based on a data base first style using stored procdures
-SecurityHelper : has a class by the same name that can staticly build
salt and hash passwords with it or rehash passwords with a stored salt.
Can't read my own database :'(
-DataBaseHelper : has a class by the same name that simply gets a static
near global database,
-Bunch of contorllers, views, and models
- notable : UtilitesControllers that handles server side data checks for
the java and the AccountController that logs in and creates sessions.

Notes: added to the web.config  for views the namespace for the ext.
methods.
I took a Database First style as I felt it was both more safe and harder
to do anyway as an example.