Quite a bit of DB changes and finish up on the data base
changes to the DB
-Projects now have a non-nullable start and end.
-Projects now have nullable actual starts and ends
-There is now a nullable Task Manager field that linkes to a new linker
table to link tasks to a Task Type Manager for the messaging system.
Table relates task types back to a manager or company so one can message
with them.
-The Messages table has had a overhall where there a alot of nullable
fields to relate it back to a project and or managment of some kind.
This is to allow for the message to be seen by the proper people.
-DB got a new create project procedure to allow for the creation of
projects.

Actual app changes
-Finished the dashboard as much as possible at this point. Layout is at
least done.
-Created a new controller with two new views for the projects. Not much
fille out in them however.